Roll back router, use pretender + convergent assertions
