--- src/routes/home.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function Home(props) {
-  return (
-    <h1>Hello World!</h1>
-  );
-}
